add service to searchpicto, fix test
